bsky: extract data classes
